+++ slice/IceRpc/Slice/DispatchException.slice
@@ -1,6 +1,6 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 
-module IceRpc;
+module IceRpc::Slice;
 
 /// A dispatch exception is thrown when a server fails to dispatch a request. Since it's built-in, any client can decode
 /// this exception.
