+++ slice/IceRpc/Slice/UnknownException.slice
@@ -1,6 +1,6 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 
-module IceRpc;
+module IceRpc::Slice;
 
 /// When a client decodes the payload of a "failure" response and does not know the type ID for the encoded exception,
 /// it throws an UnknownException.
