+++ src/IceRpc/Slice/UnknownException.cs
@@ -1,6 +1,6 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 
-namespace IceRpc;
+namespace IceRpc.Slice;
 
 /// <summary>Provides default message for class DispatchException generated by the Slice compiler.</summary>
 public partial class UnknownException
