+++ tests/IceRpc.Tests/Slice/TypeIdAttributeTests.cs
@@ -34,7 +34,7 @@ public sealed class TypeIdAttributeTests
     /// <summary>A collection of types generated from Slice definitions and its expected type IDs.</summary>
     private static readonly Dictionary<Type, string> _typeIds = new()
     {
-        [typeof(ServiceProxy)] = "::IceRpc::Service",
+        [typeof(ServiceProxy)] = "::IceRpc::Slice::Service",
         [typeof(MyClass)] = "::IceRpc::Tests::Slice::TypeIdAttributeTestNamespace::MyClass",
         [typeof(IMyInterfaceProxy)] = "::IceRpc::Tests::Slice::TypeIdAttributeTestNamespace::MyInterface",
         [typeof(MyInterfaceProxy)] = "::IceRpc::Tests::Slice::TypeIdAttributeTestNamespace::MyInterface",
@@ -51,7 +51,7 @@ public sealed class TypeIdAttributeTests
     /// <summary>A collection of types generated from Slice definitions and its expected default path.</summary>
     private static readonly Dictionary<Type, string> _defaultPaths = new()
     {
-        [typeof(ServiceProxy)] = "/IceRpc.Service",
+        [typeof(ServiceProxy)] = "/IceRpc.Slice.Service",
         [typeof(MyClass)] = "/IceRpc.Tests.Slice.TypeIdAttributeTestNamespace.MyClass",
         [typeof(IMyInterfaceProxy)] = "/IceRpc.Tests.Slice.TypeIdAttributeTestNamespace.MyInterface",
         [typeof(MyInterfaceProxy)] = "/IceRpc.Tests.Slice.TypeIdAttributeTestNamespace.MyInterface",
